enemy/chest highlighting + chest editor changes
Enemies now get highlighted on the map with an X centered on their position.
Chest now get highlighted on the map with a yellow border on their position.
Changed new/remove buttons to always show but be disabled instead when needed.
Changed chest position to hexadecimal display to fit in with the position display and highlighting.
Set gridbox selected index to -1.
Set gridbox selected index to -1 on map switch.
Commented out code that looked like it made it select a 16*16 field when index was set to -1.
Added the highlighting draws in the GridBox_Paint.