focus shift + empty list error fix + patch file
Not made in the train but whatever
Makes pressing a button that would open a window focus on that window if it is already open.
Empty lists would be 00 instead of FF due to not getting in the loop.
Can now add external patches to the Patches.event file and they will be included during build.