everything in progress, nothing done
another change dump, but this time the program is in a somewhat working state and has most of the backend work done other then yet another rework of how the object placement editor file works.

Added baseclass for changes and an extra identifier which should make it more flexible.
Text editing in its infancy, no support for the special characters yet.
Language table in constants for getting all the text.
Language manager to handle all the text entries, personal decision was made that a single new entry would be one for each language.
Attempt to add a version number to project files to allow for conversions in one place should major changes happen to changes (which will likely happen).
Same tiles highlighting for rooms, not yet a display for tile behavior in the room editor.
File association should now open the project directly.
Switch to xml for the object descriptions, should be a lot clearer to edit/add to compared to the eldritch hjson file that I cant even fully understand anymore, not yet implemented into the object editor.
Defiled the GridBoxComponent to add in "marker" which are small functions which will draw marks in the room editor or any other place using the GridBoxComponent if needed, can be added to dynamically and is planned to be used in combination with the xml files for things like marking multiple locations.
UndoRedoManager added and an example of it when editing a room, can now ctrl+z and ctrl+shift+z/ctrl+y to undo/redo.
More descriptive variable name in DrawingUtil.
Can now copy/paste larger blocks of tiles while editing rooms by holding right click, no marker made to correctly show this yet.

Hopefully with this I'll feel less like I'm keeping the project hostage as now changes should be able to happen without me saying "don't touch that part, I'm still working there"