Rework done, smaller changes from now on
DO NOT LOAD OLD PROJECTS WITH THIS!

This version should have all the features working.

What has changed

--Functional--
All saved changes are now in JSON format.
All built event files are now in a Build folder inside the project.
The Object Placement Editor now uses xml files to generate its layout for different objects and all object lists of a room can be changed. (many of the object types still need to be added)
Palette's can be changed. (first 5 entries of palette row 2 won't reflect in game as they seem to be gotten from somewhere else)
All text entries can now be changed.
Copies of the baserom are now put outside the project folder with a regional extension. (_EU, _US, _JP)
Created builds end up in a builds folder outside the project folder.
Changes to a tileset/palette are also shown in the current room without the need for a reload.
Chest editor removed as its just another list.
Key combinations like CTRL+S or CTRL+B will also activate in windows other than the room editor.
Adding of user defined xml files for use with the object editor is possible.
Undo/redo is implemented but largely unused, currently only used for tile changes and moving from room to room.
(likely more that I forgot and will appear in the release notes when its ready for release)

-- Backend --
A chunk of logic and information from Room is moved to Area.
PaletteSet and Area have their own managers now.
HJSON object filters removed as it was an unreadable mess with it all in one file, using XML instead spread among many files.
Helper for XML entry reading created.
NumberUtil created to make number string conversion simpler.
Hex extension method changed to allow padding the value.
All the drawing logic was moved to DrawingUtil to avoid duplicate code.
Added methods to DataHelper for writing and reading the JSON files for the changes.
The WarpEditorWindow now uses correct facing values and uses the new marker system.
SubWindow used as a base for all the windows other than the MainWindow.
TextConvert.xml added for use with the Text Editor.
StandardNames.txt sorted and FFFFFF's changed to -1.
All Tilesets of an Area are now loaded but can not yet be looked at with the Metatile Editor.
PaletteSet holds a bit more information for use with the editor,
LanguageManager added to hold all the banks and entries.
Language pointer added to constants, may be swapped for JP/US.
Changes are more generic (no more numbers in their names) and now have a base class depending on the scope to lessen duplicate code.
Started on the adding of rooms, this is still unused and incomplete.
A start has been made for conversion of older project to the new style but its not yet finished and not enabled, this is planned to be done for the release build.