Enable project conversion
<IF YOU HAVE MADE CHANGES TO THE BASEROM, MAKE SURE TO BACK IT UP AND POINT THE PROJECT FILE TO IT AFTER CONVERSION>
Converts the .event files of older projects into the .json files.
.event files get put in the backup folder just in case something goes wrong.
Rom is now named TMC_(region) instead of baserom_(region).
The baserom file is either moved to the baseroms folder or removed if a rom for the same region already exists.
Names get filtered out of the project file if they already exist in StandardNames.txt.
Small change to names that were cut off when changed before.